+++ app/helpers/page_title_helper.rb
@@ -23,14 +23,16 @@ module PageTitleHelper
   # @example Set the page title based on a model name
   #   page_title Post.find(1)
   #   # example translation key: 'Editing %{singular} %{id}'
+  # @example Pass along extra arguments to interpolate
+  #   page_title @post, author: @post.author.name
   # @example Show the page title
   #   <title><%= title %></title>
   #
   # @param [String, #model_name] String or ActiveModel-compliant object
   # @return [String] page title component or formatted page title
-  def page_title(new_title = nil)
+  def page_title(new_title = nil, options = {})
     if new_title
-      title_from_string_or_record_or_class(new_title).tap do |str|
+      title_from_string_or_record_or_class(new_title, options).tap do |str|
         content_for :page_title, str
       end
     elsif content_for? :page_title
@@ -66,11 +68,11 @@ module PageTitleHelper
     I18n.translate model_defaults.first, options
   end
 
-  def title_from_string_or_record_or_class(object)
+  def title_from_string_or_record_or_class(object, options)
     if object.respond_to?(:model_name)
-      title_for_class(object)
+      title_for_class(object, options)
     elsif object.class.respond_to?(:model_name)
-      title_for_class(object.class, id: object.to_param)
+      title_for_class(object.class, options.reverse_merge(id: object.to_param))
     else
       object
     end